+++ R/update.meta.R
@@ -491,7 +491,8 @@ update.meta <- function(object,
     object$subset <- NULL
     ##
     object$data <- data.frame(.studlab = object$studlab,
-                              .exclude = rep_len(FALSE, length(object$studlab)))
+                              .exclude = rep_len(FALSE,
+                                                 length(object$studlab)))
     ##
     if (!is.null(object$byvar))
       object$data$.byvar <- object$byvar
@@ -596,13 +597,13 @@ update.meta <- function(object,
   missing.id <- missing(id)
   ##
   if (!missing.id)
-    id <- eval(mf[[match("id", names(mf))]],
-               data, enclos = sys.frame(sys.parent()))
+    ...id <- eval(mf[[match("id", names(mf))]],
+                  data, enclos = sys.frame(sys.parent()))
   else {
     if (isCol(object$data, ".id"))
-      id <- object$data$.id
+      ...id <- object$data$.id
     else
-      id <- NULL
+      ...id <- NULL
   }
   ##
   ## Catch argument 'incr'
@@ -660,14 +661,17 @@ update.meta <- function(object,
     if (sm == "ASD") {
       if (!missing.incr && any(incr != 0))
         warning("Note, no continuity correction considered for ",
-                "arcsine difference (sm = \"ASD\").")
+                "arcsine difference (sm = \"ASD\").",
+                call. = FALSE)
       incr <- 0
       object$data$.incr <- 0
     }
     ##
     if (method == "Peto") {
       if (!missing.incr && any(incr != 0))
-        warning("Note, no continuity correction considered for method = \"Peto\".")
+        warning("Note, no continuity correction considered for ",
+                "method = \"Peto\".",
+                call. = FALSE)
       incr <- 0
       object$data$.incr <- 0
     }
@@ -685,7 +689,8 @@ update.meta <- function(object,
                  method = method,
                  sm = ifelse(method == "GLMM", "OR", sm),
                  incr = incr,
-                 allincr = allincr, addincr = addincr, allstudies = allstudies,
+                 allincr = allincr, addincr = addincr,
+                 allstudies = allstudies,
                  MH.exact = MH.exact, RR.Cochrane = RR.Cochrane,
                  Q.Cochrane = Q.Cochrane, model.glmm = model.glmm,
                  ##
@@ -734,12 +739,13 @@ update.meta <- function(object,
                   ##
                   studlab = studlab,
                   exclude = exclude,
-                  id = id,
+                  id = ...id,
                   ##
                   data = data, subset = subset,
                   ##
                   sm = sm, pooledvar = pooledvar,
-                  method.smd = method.smd, sd.glass = sd.glass, exact.smd = exact.smd,
+                  method.smd = method.smd, sd.glass = sd.glass,
+                  exact.smd = exact.smd,
                   ##
                   method.ci = ifelse(is.null(method.ci), gs("method.ci.cont"),
                                      method.ci),
@@ -831,7 +837,7 @@ update.meta <- function(object,
                  ##
                  studlab = studlab,
                  exclude = exclude,
-                 id = id,
+                 id = ...id,
                  ##
                  data = data.m, subset = subset,
                  ##